refactor: added async games; refactor: broke out classes to modules; feat: created API; refactor: refactored player attribute assignment; refactor: refactored playoff function
