+++ config.json
@@ -1,6 +1,6 @@
 {
     "leagueConfig": {
-        "seasons": 4,
+        "seasons": 1,
         "deleteDataOnRestart": true,
         "freeAgencyRounds": 2
     },
@@ -20,13 +20,13 @@
         "Denver Northern Lights",
         "Seattle Cranes",
         "Cleveland Rocks",
-        "Washington Belts",
+        "Washington Companions",
         "Tampa Bay Bees",
         "Miami Jetskis",
         "Toronto Imaginaries",
         "San Diego Sand Dollars",
         "Phoenix Dry Heat",
-        "Las Vegas Residencies"
+        "Las Vegas Residents"
 
     ],
     "players": [
@@ -43,7 +43,6 @@
         "Omar Dominguez",
         "Juana Palmer",
         "Manuela Huerta",
-        "Donovan Santos",
         "Bizzy Bone",
         "Large Ant",
         "Kevin Oddly",
@@ -52,7 +51,7 @@
         "Sister Lettuce",
         "Coco Mattel",
         "Chi Chi Eichelberger",
-        "Kangaroo DriveIn",
+        "Kangaroo Drive-In",
         "Roid Drill",
         "Anger Brutal",
         "Rusty Mateo",
@@ -69,37 +68,16 @@
         "Katie Phoenix",
         "Krista Button",
         "Ellie Living",
-        "Alexia Joyce",
         "Junior Man",
-        "Marcus Duff",
         "Lester Carroll",
-        "Frank Foster",
         "Stanislaw Dohrman",
         "Thibaut Swathi",
         "Washington Routledge",
-        "Chadwick Pena",
-        "Bernard Sweeney",
         "Misha Duskin",
-        "Sergei Berezin",
-        "Tasha Sorokin",
-        "Yulia Popoff",
-        "Kiril Popoff",
-        "Grigori Gurin",
         "Seno Tomoyuki",
         "Hashira Kanzaburo",
         "Yoshihara Yasutake",
         "Shinobu Kane",
-        "Diedrick Hoffner",
-        "Bernard Schleicher",
-        "Mubarak el-Basa",
-        "Najeema al-Shariff",
-        "Suhail al-Laham",
-        "Yi Jang",
-        "Xue Tam",
-        "An Soo Hoo",
-        "Lin Meng",
-        "Shen Jeung",
-        "Hyon Jin-Hee",
         "Winston Jazz Hands",
         "Loving Spectator",
         "Golden Prince",
@@ -110,9 +88,7 @@
         "Trinity Pantsdown",
         "Chicago Houston",
         "Johnny Steele",
-        "Thelma Wallace",
         "Minnie Hester",
-        "Dianne Armstrong",
         "Thunder Princess",
         "Velocity Man",
         "Fast Sweetness",
@@ -129,19 +105,12 @@
         "Jimmy Mars",
         "Sam Colt",
         "Alexia Legend",
-        "Zach Mythic",
-        "Bob Sinclair",
         "Indiana Good",
         "Sandie Pip",
         "Sigi Allison",
         "Ignacious Pinkerton",
         "Chad Kingfish",
-        "Vincent Castillo",
-        "Clifford Wolfe",
-        "Emerson Swanson",
-        "Wesley Nash",
         "Hershel Pruitt",
-        "Domingo Hood",
         "Kitty Jacobson",
         "Vivian Leblanc",
         "Griffin Star",
@@ -158,8 +127,6 @@
         "Rachel Chachki",
         "Pearl Miller",
         "Alaska Markova",
-        "Alan MacLoughlin",
-        "Callum Laverty",
         "Callaway Casey",
         "Ocean Emerson",
         "Trowbridge Langdon",
@@ -209,8 +176,6 @@
         "Hazelmere Drach",
         "Jasper Redwood",
         "Malachi Brock",
-        "Darryl Lawson",
-        "Tommy Mason",
         "Mario Hendrix",
         "Fernando Harding",
         "Wesley Sawyer",
@@ -220,16 +185,12 @@
         "Hatano Akimoro",
         "Sugi Suemoro",
         "Miura Hidenaka",
-        "Okabe Kiyoemon",
         "Maki Sada",
         "Alfie MacGilloway",
         "Fionn Kerrane",
         "Kayden O'Boland",
-        "Éabha Canny",
+        "Eabha Canny",
         "Willow O'Hanlon",
-        "Cahit Gardezi",
-        "Nabeel Mahar",
-        "Parisuz Khosa",
         "Dominique Gustave",
         "Thony Jacques",
         "Herold Louis Jeune",
@@ -239,29 +200,12 @@
         "Lieze Peleriaux",
         "Inne Blomme",
         "Toon Van Laethem",
-        "Cosmo Reed",
         "Pierre Younger",
         "Leighton Buckingham",
-        "Si Jae-Sun",
-        "Kung Jeong-Mun",
-        "Tang Song-Jin",
-        "Ki Jin-Yung",
-        "Won Hyo-Ju",
-        "Hassan Baz",
-        "Osman Awad",
-        "Shaaban Gerges",
-        "Khaled Essa",
         "Ariki Taimana",
         "Tama Hamutana",
         "Hohepa Wikiriwhi",
-        "Mokihi Ropata",
-        "Rolando Costa",
-        "Sebastian Molina",
-        "Aakash Roychaudhuri",
-        "Sumit Havaldar",
         "Ragunath Jogalekar",
-        "Siddhi Gazdar",
-        "Undead Emperor",
         "Doctor Small Queen",
         "Loaded Gun",
         "Rich Flavor",
@@ -269,7 +213,16 @@
         "TeeTee Moonshine",
         "Bud 'Lite' Rainwater",
         "Generic Mann",
-        "Ultimate Android"
+        "Ultimate Android",
+        "SuperFootballTron-01",
+        "SuperFootballTron-02",
+        "SuperFootballTron-03",
+        "SuperFootballTron-04",
+        "SuperFootballTron-07",
+        "UltraFootballTron-62",
+        "Zealous Tiger",
+        "Donnie Fame",
+        "Frau Robin"
